+++ requests_retry_session/requests_retry_session.py
@@ -35,7 +35,8 @@ Created on Nov 2, 2020
 """
 
 # Because we want to support Python 3.6 and 3.9, use old-style type hint syntax
-from typing import Tuple, Union
+from typing import Optional, Tuple
+from typing_extensions import TypedDict, Unpack
 
 import requests
 
@@ -50,9 +51,16 @@ DEFAULT_READ_TIMEOUT = 10
 DEFAULT_RETRIES = 10,
 DEFAULT_STATUS_FORCELIST = (500, 502, 503, 504)
 
+class RequestsRetryAdapterArgs(TypedDict, total=False):
+    retries: int
+    backoff_factor: float
+    status_forcelist: Tuple[int, ...]
+    connect_timeout: int
+    read_timeout: int
+
 
 def requests_session(adapter: requests.adapters.HTTPAdapter,
-                     session: Union[None, requests.Session] = None,
+                     session: Optional[requests.Session] = None,
                      protocol: str = DEFAULT_PROTOCOL) -> requests.Session:
     session = session or requests.Session()
     # Must mount to http://
@@ -76,16 +84,8 @@ def requests_retry_adapter(retries: int = DEFAULT_RETRIES,
     return TimeoutHTTPAdapter(max_retries=retry, timeout=(connect_timeout, read_timeout))
 
 
-def requests_retry_session(retries: int = DEFAULT_RETRIES,
-                           backoff_factor: float = DEFAULT_BACKOFF_FACTOR,
-                           status_forcelist: Tuple[int, ...] = DEFAULT_STATUS_FORCELIST,
-                           connect_timeout: int = DEFAULT_CONNECT_TIMEOUT,
-                           read_timeout: int = DEFAULT_READ_TIMEOUT,
-                           session: Union[None, requests.Session] = None,
-                           protocol: str = DEFAULT_PROTOCOL) -> requests.Session:
-    adapter = requests_retry_adapter(retries=retries,
-                                     backoff_factor=backoff_factor,
-                                     status_forcelist=status_forcelist,
-                                     connect_timeout=connect_timeout,
-                                     read_timeout=read_timeout)
+def requests_retry_session(session: Optional[requests.Session] = None,
+                           protocol: str = DEFAULT_PROTOCOL,
+                           **adapter_kwargs: Unpack[RequestsRetryAdapterArgs]) -> requests.Session:
+    adapter = requests_retry_adapter(**adapter_kwargs)
     return requests_session(adapter=adapter, session=session, protocol=protocol)